Sacado/Stokhos:  Put create_mirror() overload declarations in separate file
This allows for the proper include order when including multiple
Sacado/Stokhos scalar types in the same file, since all of the
create_mirror() overloads need to be declared before
Kokkos_CopyViews.hpp is included.  Fixes build errors in Stokhos and
TrilinosCouplings for CUDA when UVM is disabled.
