+++ packages/stokhos/src/sacado/kokkos/vector/Sacado_MP_Vector.hpp
@@ -59,6 +59,7 @@
 #include "Stokhos_mpl_for_each.hpp"
 #include "Stokhos_MemoryTraits.hpp"
 #include "Stokhos_Is_Constant.hpp"
+#include "Stokhos_ViewStorage.hpp"
 
 #include "Kokkos_View_Utils.hpp"
